requirements.txt
streamlit
janome
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
google-cloud-translate
alkana